al eliminar productos muestra un msg de que si esta seguro.

+++ Proyecto Final/src/components/CrudAdmin/productos/CadaCarro/Chevrolet.jsx
@@ -16,15 +16,20 @@ function ChevroletMostrar() {
   }, []);
 
   const eliminarChevrolet = async (id) => {
-    deleteChevrolet(id)
-    setTimeout(() => {
-      datos()
-    }, 100)
+    if (confirm("Estas seguro que deseas eliminar este producto??") == true) {
+      deleteChevrolet(id)
+    } else {
+      return false
+    }
+    datos();
   }
-
   const enviarDatos = async () => {
-    await putChevrolet( imageUrl, Year, Price, Chevrolet, id);
-    setModalAbierto(false);
+    if (imageUrl.trim() === '' || Year.trim() === '' || Price.trim() === '') {
+      alert("No puede dejar campos vacios!")
+    } else {
+      await putChevrolet(imageUrl, Year, Price, Chevrolet, id);
+      setModalAbierto(false);
+    }
     datos();
   };
 
@@ -34,7 +39,7 @@ function ChevroletMostrar() {
     setCarros(datosChevrolet);
   };
   const BotonEditar = (item) => {
-    setID(item.id); 
+    setID(item.id);
     setImageUrl(item.Url);
     setYear(item.año);
     setPrice(item.precio);
@@ -56,16 +61,16 @@ function ChevroletMostrar() {
               <img className="imgCarros" src={item.Url} alt="" />
             </div>
             <div className="divDescription">
-            <div className="div2Carros">
+              <div className="div2Carros">
                 <h6 className="h6Txt">{item.marca}</h6>
                 <p className="precio">{item.año}</p>
                 <p className="precio2">{"$" + item.precio}</p>
               </div>
               <div>
-              <div className="divInpCrud">
-            <input className="Inpdelete" onClick={() => eliminarChevrolet(item.id)} type="button" value="Eliminar" />
-            <input className="Editar" onClick={() => BotonEditar(item)} type="button" value="Editar" />
-            </div>
+                <div className="divInpCrud">
+                  <input className="Inpdelete" onClick={() => eliminarChevrolet(item.id)} type="button" value="Eliminar" />
+                  <input className="Editar" onClick={() => BotonEditar(item)} type="button" value="Editar" />
+                </div>
               </div>
             </div>
 
@@ -90,7 +95,7 @@ function ChevroletMostrar() {
                 <div className="mini2">
                   <input className="cerrarModal" onClick={closeModal} type="button" value="X" />
                   <div>
-                  <option  className='inpPost' value={Chevrolet} onChange={(e) => setChevrolet(e.target.value)}>Chevrolet</option>
+                    <option className='inpPost' value={Chevrolet} onChange={(e) => setChevrolet(e.target.value)}>Chevrolet</option>
                     <input className="inpPost" type="text" placeholder="URL de la Imagen" value={imageUrl} onChange={handleInputChange} />
                     <input className="inpPost" type="text" placeholder="Year" value={Year} onChange={(e) => setYear(e.target.value)} />
                     <input className="inpPost" type="text" placeholder="Precio en USD" value={Price} onChange={(e) => setPrice(e.target.value)}
