+++ Proyecto Final/src/components/CrudAdmin/productos/CadaCarro/Honda.jsx
@@ -19,21 +19,27 @@ function HondaMostrar() {
     setCarros(datosHonda);
   };
   const eliminarHonda = async (id) => {
-    deleteHonda(id)
-    setTimeout(() => {
-      datos()
-    }, 100)
+    if (confirm("Estas seguro que deseas eliminar este producto??") == true) {
+      deleteHonda(id)
+    } else {
+      return false
+    }
+    datos();
   }
   const BotonEditar = (item) => {
-    setID(item.id); 
+    setID(item.id);
     setImageUrl(item.Url);
     setYear(item.año);
     setPrice(item.precio);
     setModalAbierto(true);
   };
   const enviarDatos = async () => {
-    await putHonda( imageUrl, Year, Price, Honda, id);
-    setModalAbierto(false);
+    if (imageUrl.trim() === '' || Year.trim() === '' || Price.trim() === '') {
+      alert("No puede dejar campos vacios!")
+    } else {
+      await putHonda(imageUrl, Year, Price, Honda, id);
+      setModalAbierto(false);
+    }
     datos();
   };
   const closeModal = () => {
@@ -53,17 +59,17 @@ function HondaMostrar() {
               <img className="imgCarros" src={item.Url} alt="" />
             </div>
             <div className="divDescription">
-            <div className="div2Carros">
+              <div className="div2Carros">
                 <h6 className="h6Txt">{item.marca}</h6>
                 <p className="precio">{item.año}</p>
                 <p className="precio2">{"$" + item.precio}</p>
               </div>
-            <div>
-            <div className="divInpCrud">
-            <input className="Inpdelete" onClick={() => eliminarHonda(item.id)} type="button" value="Eliminar" />
-            <input className="Editar" onClick={() => BotonEditar(item)} type="button" value="Editar" />
-            </div>
-            </div>
+              <div>
+                <div className="divInpCrud">
+                  <input className="Inpdelete" onClick={() => eliminarHonda(item.id)} type="button" value="Eliminar" />
+                  <input className="Editar" onClick={() => BotonEditar(item)} type="button" value="Editar" />
+                </div>
+              </div>
             </div>
           </div>
         ))}
@@ -86,7 +92,7 @@ function HondaMostrar() {
                 <div className="mini2">
                   <input className="cerrarModal" onClick={closeModal} type="button" value="X" />
                   <div>
-                  <option  className='inpPost' value={Honda} onChange={(e) => setHonda(e.target.value)}>Honda</option>
+                    <option className='inpPost' value={Honda} onChange={(e) => setHonda(e.target.value)}>Honda</option>
                     <input className="inpPost" type="text" placeholder="URL de la Imagen" value={imageUrl} onChange={handleInputChange} />
                     <input className="inpPost" type="text" placeholder="Year" value={Year} onChange={(e) => setYear(e.target.value)} />
                     <input className="inpPost" type="text" placeholder="Precio en USD" value={Price} onChange={(e) => setPrice(e.target.value)}
